+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Fixed
+
+- Update `Baggage` in `go.opentelemetry.io/otel/propagation` and `Parse` in `go.opentelemetry.io/otel/baggage` to comply with W3C Baggage specification limits. (#7880)
+
 <!-- Released section -->
 <!-- Don't change this section unless doing release -->